+++ src/test/kotlin/tornadofx/testapps/DataGridTest.kt
@@ -1,11 +1,8 @@
 package tornadofx.testapps
 
-import javafx.beans.binding.Bindings
-import javafx.collections.FXCollections
 import javafx.scene.control.SelectionMode
 import tornadofx.*
 import tornadofx.testapps.DataGridTestApp.Companion.images
-import java.util.concurrent.Callable
 
 class DataGridTestApp : App(DataGridTest::class, DataGridStyles::class) {
     companion object {
@@ -40,13 +37,13 @@ class DataGridTest : View("DataGrid") {
                 combobox(values = images.keys.toList()) {
                     promptText = "Select images"
                     valueProperty().onChange {
-                        datagrid.items = FXCollections.observableArrayList(images[it])
+                        datagrid.items.setAll(images[it])
                     }
+                    shortcut("k") { value = "kittens" }
+                    shortcut("p") { value = "puppies" }
                 }
-                button("Add") {
-                    setOnAction {
-                        datagrid.items.add("http://i.imgur.com/bvqTBT0b.jpg")
-                    }
+                button("Add").action {
+                    datagrid.items.add("http://i.imgur.com/bvqTBT0b.jpg")
                 }
             }
         }
@@ -55,8 +52,8 @@ class DataGridTest : View("DataGrid") {
                 setPrefSize(550.0, 550.0)
 
                 selectionModel.selectionMode = SelectionMode.MULTIPLE
-
-                maxCellsInRow = 3
+                //maxCellsInRow = 3
+                //maxRows = 3
                 cellWidth = 164.0
                 cellHeight = 164.0
 
@@ -65,12 +62,12 @@ class DataGridTest : View("DataGrid") {
                 }
 
                 onUserSelect(1) {
-                    println("Selected $selectedItem")
+                    println("Selected $it")
                 }
             }
         }
         bottom {
-            label(Bindings.createStringBinding(Callable {datagrid.selectionModel.selectedItems.joinToString(", ")}, datagrid.selectionModel.selectedItems))
+            label(stringBinding(datagrid.selectionModel.selectedItems) { joinToString(", ") })
         }
     }